requirements.txt
# BlisterVision Elite — Python Dependencies
# Install: pip install -r requirements.txt

# Core ML
ultralytics>=8.2.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0.0

# Data & utils
pyyaml>=6.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Optional: Roboflow dataset download
# roboflow>=1.1.0
